requirements.txt
annotated-types==0.7.0
anyio==4.8.0
Authlib==1.3.1
certifi==2025.1.31
cffi==1.17.1
cryptography==44.0.2
exceptiongroup==1.2.2
grpcio==1.70.0
grpcio-health-checking==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
protobuf==5.29.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
sniffio==1.3.1
typing_extensions==4.12.2
validators==0.34.0
weaviate-client==4.11.1
